Refactor sign-in process and add image upload utility;
